requirements.txt
celery[redis]
chardet
flask
guess-language
hiredis
pycountry
mwoauth
youtube-dl
https://github.com/senko/python-video-converter/zipball/master#egg=videoconverter
https://github.com/wikimedia/pywikibot-core/zipball/master#egg=pywikibot
